+++ EstimoteSDK/EstimoteSDK.framework/Versions/A/Headers/ESTRequestBeaconMac.h
@@ -1,0 +1,57 @@
+//
+//   ______     _   _                 _          _____ _____  _  __
+//  |  ____|   | | (_)               | |        / ____|  __ \| |/ /
+//  | |__   ___| |_ _ _ __ ___   ___ | |_ ___  | (___ | |  | | ' /
+//  |  __| / __| __| | '_ ` _ \ / _ \| __/ _ \  \___ \| |  | |  <
+//  | |____\__ \ |_| | | | | | | (_) | ||  __/  ____) | |__| | . \
+//  |______|___/\__|_|_| |_| |_|\___/ \__\___| |_____/|_____/|_|\_\
+//
+//
+//  Copyright (c) 2015 Estimote. All rights reserved.
+
+#import "ESTRequestGetJSON.h"
+#import <CoreLocation/CoreLocation.h>
+
+#define ESTRequestBeaconMacErrorDomain @"ESTRequestBeaconMacErrorDomain"
+
+/**
+ *  Errors occurring for request.
+ */
+typedef NS_ENUM(NSInteger, ESTRequestBeaconMacError)
+{
+    /**
+     *  Unknown error occurred.
+     */
+    ESTRequestBeaconMacErrorUnknown
+};
+
+typedef void(^ESTRequestBeaconMacBlock)(NSString *macAddress, NSError *error);
+
+/**
+ *  ESTRequestBeaconMac allows to get MAC address for beacon based on CLBeacon
+ *  (ProximityUUID, Major, Minor) identification. Method is publicly available
+ *  for all beacons.
+ */
+
+@interface ESTRequestBeaconMac : ESTRequestGetJSON
+
+@property (nonatomic, strong, readonly) CLBeacon *beacon;
+
+
+/**
+ *  Initialise request with beacon.
+ *
+ *  @param beacon beacon should be used to get MAC address.
+ *
+ *  @return instance of request
+ */
+- (instancetype)initWithBeacon:(CLBeacon *)beacon;
+
+/**
+ *  Methods allows to send request with completion block invoked as a result.
+ *
+ *  @param completion Completion block with returned data (CLLocation object).
+ */
+- (void)sendRequestWithCompletion:(ESTRequestBeaconMacBlock)completion;
+
+@end