+++ EstimoteSDK/EstimoteSDK.framework/Versions/A/Headers/ESTRequestPatchJSON.h
@@ -1,0 +1,30 @@
+//
+//   ______     _   _                 _          _____ _____  _  __
+//  |  ____|   | | (_)               | |        / ____|  __ \| |/ /
+//  | |__   ___| |_ _ _ __ ___   ___ | |_ ___  | (___ | |  | | ' /
+//  |  __| / __| __| | '_ ` _ \ / _ \| __/ _ \  \___ \| |  | |  <
+//  | |____\__ \ |_| | | | | | | (_) | ||  __/  ____) | |__| | . \
+//  |______|___/\__|_|_| |_| |_|\___/ \__\___| |_____/|_____/|_|\_\
+//
+//
+//  Copyright (c) 2015 Estimote. All rights reserved.
+
+#import <Foundation/Foundation.h>
+#import "ESTRequestBase.h"
+
+#define ESTRequestPatchJSONErrorDomain @"ESTRequestPatchJSONErrorDomain"
+
+typedef NS_ENUM(NSInteger, ESTRequestPatchJSONError)
+{
+    ESTRequestPatchJSONErrorNoData,
+    ESTRequestPatchJSONErrorNotAuthorized,
+    ESTRequestPatchJSONErrorNotResource,
+    ESTRequestPatchJSONErrorCloudReturnedError
+};
+
+
+@interface ESTRequestPatchJSON : ESTRequestBase
+
+- (void)setParams:(NSDictionary *)params forRequest:(NSMutableURLRequest *)request;
+
+@end